// src/core/models/enums.ts
export enum QuizType {
  GAME = 'GAME',
  PRACTICE_SET = 'PRACTICE_SET'
}
export enum QuestionType {
  MCQ = 'MCQ',
  MSQ = 'MSQ',
  TrueFalse = 'TRUE_FALSE',
  // Poll = 'POLL',
  // FillInTheBlank = 'FILL_IN_THE_BLANK',
  // MatchTheFollowing = 'MATCH_THE_FOLLOWING',
}
export enum MediaType {
  HTML = 'HTML',
  IMAGE = 'IMAGE',
  VIDEO = 'VIDEO',
  YOUTUBE = 'YOUTUBE',
  LATEX = 'LATEX'
}